requirements.txt
streamlit
openai
pypdf
python-dotenv
requests
pandas
